Disable Chrome's auto translation feature.
The actual issue seems to be that there's not enough "real" text on
the page, but that's not something I can currently solve.

I don't like this solution incase non-English speakers want to use
the site, but I also can't have the majority of Chattanooga being
asked if they want translate the page from Spanish every page load.